rename ontology package to bigquery_ontology
